requirements.txt
bleach==6.2.0
bokeh==3.7.3
certifi==2025.8.3
charset-normalizer==3.4.2
contourpy==1.3.3
idna==3.10
Jinja2==3.1.6
linkify-it-py==2.0.3
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
narwhals==2.0.1
numpy==2.3.2
packaging==25.0
pandas==2.3.1
panel==1.7.5
param==2.2.1
pillow==11.3.0
psycopg2-binary==2.9.10
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyviz_comms==3.0.6
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.5.0
webencodings==0.5.1
xyzservices==2025.4.0
